+++ apps/backend/server.ts
@@ -4,7 +4,6 @@ import { Server, Socket } from "socket.io";
 import cors from "cors";
 import bodyParser from "body-parser";
 
-
 const app = express();
 const server = http.createServer(app);
 const io = new Server(server, {
@@ -14,7 +13,6 @@ const io = new Server(server, {
 app.use(cors({ origin: "*" }));
 app.use(bodyParser.urlencoded({ extended: true }));
 app.use(bodyParser.json());
-
 
 // Start the server on port 3001
 const PORT = process.env.PORT || 3001;