+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 1. Move frontend/backend to `apps` directory which is standard for js monorepos
 1. Install turborepo and move to pnpm workspaces
    1. Backend: install tsx to run server
+1. Install chokidar to run the server with hot-reload